Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/module/inline-async-execorder-expected.txt

CONSOLE MESSAGE: line 1: ReferenceError: Can't find variable: loaded
CONSOLE MESSAGE: line 1: ReferenceError: Can't find variable: loaded
CONSOLE MESSAGE: line 3: ReferenceError: Can't find variable: loaded
CONSOLE MESSAGE: line 3: ReferenceError: Can't find variable: loaded

Harness Error (FAIL), message = ReferenceError: Can't find variable: loaded

FAIL Inline async module script execution order assert_array_equals: lengths differ, expected 6 got 0
